requirements.txt
selenium~=4.23.1
pytest~=8.2.2
allure-python-commons~=2.13.5